requirements.txt
httpx
pydantic
python-dotenv
requests
urllib3
